Trait trowel::flipped::DrawTargetExt2

pub trait DrawTargetExt2: DrawTarget + Sized {
    fn flipped(&mut self, axes: Axes) -> Flipped<'_, Self>;
}

Required Methods

fn flipped(&mut self, axes: Axes) -> Flipped<'_, Self>

Creates a translated draw target based on this draw target.

All drawing operations are translated by offset pixels, before being passed to the parent draw target.

Examples

use embedded_graphics::{
    mock_display::MockDisplay,
    mono_font::{ascii::FONT_6X9, MonoTextStyle},
    pixelcolor::BinaryColor,
    prelude::*,
    text::Text,
};

let mut display = MockDisplay::new();
let mut translated_display = display.translated(Point::new(5, 10));

let style = MonoTextStyle::new(&FONT_6X9, BinaryColor::On);

// Draws text at position (5, 10) in the display coordinate system
Text::new("Text", Point::zero(), style).draw(&mut translated_display)?;